Vehicle Tracking/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
ultralytics
Pillow
numpy
faiss-cpu
pyyaml
torchreid
